% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basics_annuities.R
\name{calc_compound_discount}
\alias{calc_compound_discount}
\title{Calculate Compound Discount (Present Value)}
\usage{
calc_compound_discount(S, i, t)
}
\arguments{
\item{S}{Numeric. The nominal value.}

\item{i}{Numeric. The annual interest rate.}

\item{t}{Numeric. The time in years.}
}
\value{
Numeric. The present value A.
}
\description{
Calculates the present value A in compound interest.
Formula: A = S * (1 + i)^(-t)
}
