% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_investment_analysis.R
\name{payback_period}
\alias{payback_period}
\title{Calculate Payback Period}
\usage{
payback_period(cash_flows, times = NULL, i = 0)
}
\arguments{
\item{cash_flows}{Numeric vector. The series of cash flows.}

\item{times}{Numeric vector. The times corresponding to cash flows.}

\item{i}{Numeric. Discount rate (set 0 for simple payback, >0 for discounted).}
}
\value{
Numeric. The payback period (interpolated).
}
\description{
Calculates the (simple and discounted) payback period.
}
