% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_portfolio_selection.R
\name{portfolio_return_two_assets}
\alias{portfolio_return_two_assets}
\title{Two Asset Portfolio Return}
\usage{
portfolio_return_two_assets(w, mu1, mu2)
}
\arguments{
\item{w}{Numeric. Weight of asset 1.}

\item{mu1}{Numeric. Expected return of asset 1.}

\item{mu2}{Numeric. Expected return of asset 2.}
}
\value{
Numeric. Portfolio expected return.
}
\description{
Calculates expected return of a portfolio of two assets.
Formula: mu_p = w * mu1 + (1-w) * mu2
}
