% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_formulas.R
\name{show_formula}
\alias{show_formula}
\title{Recall a Formula}
\usage{
show_formula(name)
}
\arguments{
\item{name}{A string specifying the name of the formula.
Matches are performed using loose keywords.
Run `show_formula()` without arguments to see the list of available formulas.}
}
\value{
Prints the formula text to the console.
}
\description{
This function prints a requested formula or concept to the console.
}
\examples{
show_formula("compound interest")
show_formula("gordon")
}
