% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_proofs.R
\name{show_proof}
\alias{show_proof}
\title{Recall a Proof}
\usage{
show_proof(name)
}
\arguments{
\item{name}{A string specifying the name of the proof. 
Matches are performed using loose keywords (e.g., "annual rate", "put call").
Run `show_proof()` without arguments to see the list of available proofs.}
}
\value{
Prints the proof text to the console.
}
\description{
This function prints a requested proof or formula derivation to the console.
}
\examples{
show_proof("put call")
}
