% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_statistics.R
\name{show_stats}
\alias{show_stats}
\title{Recall a Statistics/Regression Concept}
\usage{
show_stats(name)
}
\arguments{
\item{name}{A string specifying the name of the concept.
Matches are performed using loose keywords.
Run `show_stats()` without arguments to see the list of available topics.}
}
\value{
Prints the concept text to the console.
}
\description{
This function prints a requested statistics or regression concept to the console.
}
\examples{
show_stats("ols")
show_stats("r squared")
}
