% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_advanced_investments.R
\name{profitability_index}
\alias{profitability_index}
\title{Profitability Index}
\usage{
profitability_index(cash_flows, times = NULL, i)
}
\arguments{
\item{cash_flows}{Numeric vector. The cash flows.}

\item{times}{Numeric vector. The times (default 0:(n-1)).}

\item{i}{Numeric. The discount rate.}
}
\value{
Numeric. The Profitability Index.
}
\description{
Calculates the Profitability Index (PI).
Formula: PI = PV(Future Cash Flows) / Absolute(Initial Investment)
         PI = (NPV + |C0|) / |C0|
}
